pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "intellapi"
version = "0.1.0"
description = "AI-powered API documentation generator — analyze your backend, generate professional docs."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">= 3.10"
authors = [
    {name = "Intellapi Contributors"},
]
keywords = ["api", "documentation", "ai", "llm", "fastapi", "express", "openapi"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Documentation",
    "Topic :: Software Development :: Documentation",
]
dependencies = [
    "typer[all]>=0.9",
    "rich>=13.0",
    "pydantic>=2.0",
    "keyring>=25.0",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "httpx>=0.27",
    "anthropic>=0.40",
    "boto3>=1.35",
    "tree-sitter>=0.21",
    "tree-sitter-javascript>=0.21",
    "tree-sitter-typescript>=0.21",
    "pathspec>=0.11",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.4",
]

[project.scripts]
intellapi = "intellapi.cli:app"

[project.urls]
Homepage = "https://github.com/intellapi/intellapi"
Documentation = "https://github.com/intellapi/intellapi#readme"
Repository = "https://github.com/intellapi/intellapi"
Issues = "https://github.com/intellapi/intellapi/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/intellapi"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
